Support JsonExtensionData in STJ source-gen
