+++ src/libraries/System.Text.Json/gen/Reflection/TypeExtensions.cs
@@ -4,6 +4,7 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using System.Text.Json.SourceGeneration;
 
 namespace System.Text.Json.Reflection
 {
