+++ src/libraries/System.Text.Json/gen/TypeGenerationSpec.cs
@@ -58,6 +58,10 @@ namespace System.Text.Json.SourceGeneration
 
         public TypeGenerationSpec? NullableUnderlyingTypeMetadata { get; private set; }
 
+        public string? RuntimeTypeRef { get; private set; }
+
+        public TypeGenerationSpec? ExtensionDataPropertyTypeSpec {  get; private set; }
+
         public string? ConverterInstantiationLogic { get; private set; }
 
         // Only generate certain helper methods if necessary.
@@ -109,6 +113,8 @@ namespace System.Text.Json.SourceGeneration
             TypeGenerationSpec? collectionValueTypeMetadata,
             ObjectConstructionStrategy constructionStrategy,
             TypeGenerationSpec? nullableUnderlyingTypeMetadata,
+            string? runtimeTypeRef,
+            TypeGenerationSpec? extensionDataPropertyTypeSpec,
             string? converterInstantiationLogic,
             bool implementsIJsonOnSerialized,
             bool implementsIJsonOnSerializing,
@@ -130,6 +136,8 @@ namespace System.Text.Json.SourceGeneration
             CollectionValueTypeMetadata = collectionValueTypeMetadata;
             ConstructionStrategy = constructionStrategy;
             NullableUnderlyingTypeMetadata = nullableUnderlyingTypeMetadata;
+            RuntimeTypeRef = runtimeTypeRef;
+            ExtensionDataPropertyTypeSpec = extensionDataPropertyTypeSpec;
             ConverterInstantiationLogic = converterInstantiationLogic;
             ImplementsIJsonOnSerialized = implementsIJsonOnSerialized;
             ImplementsIJsonOnSerializing = implementsIJsonOnSerializing;
@@ -221,6 +229,11 @@ ReturnFalse:
         {
             if (ClassType == ClassType.Object)
             {
+                if (ExtensionDataPropertyTypeSpec != null)
+                {
+                    return false;
+                }
+
                 foreach (PropertyGenerationSpec property in PropertyGenSpecList)
                 {
                     if (property.TypeGenerationSpec.Type.IsObjectType() ||
