+++ src/libraries/System.Text.Json/tests/Common/TestClasses/TestClasses.Constructor.cs
@@ -923,7 +923,7 @@ namespace System.Text.Json.Serialization.Tests
         public int Int { get; set; }
     }
 
-    public class ClassWithConstructor_SimpleAndComplexParameters : ITestClassWithParameterizedCtor
+    public class ObjWCtorMixedParams : ITestClassWithParameterizedCtor
     {
         public byte MyByte { get; }
         public sbyte MySByte { get; set; }
@@ -961,7 +961,7 @@ namespace System.Text.Json.Serialization.Tests
         public ImmutableSortedSet<string> MyStringImmutableSortedSetT { get; }
         public List<string> MyListOfNullString { get; }
 
-        public ClassWithConstructor_SimpleAndComplexParameters(
+        public ObjWCtorMixedParams(
             byte myByte,
             char myChar,
             string myString,
@@ -1011,8 +1011,8 @@ namespace System.Text.Json.Serialization.Tests
             MyListOfNullString = myListOfNullString;
         }
 
-        public static ClassWithConstructor_SimpleAndComplexParameters GetInstance() =>
-            JsonSerializer.Deserialize<ClassWithConstructor_SimpleAndComplexParameters>(s_json);
+        public static ObjWCtorMixedParams GetInstance() =>
+            JsonSerializer.Deserialize<ObjWCtorMixedParams>(s_json);
 
         public static string s_json => $"{{{s_partialJson1},{s_partialJson2}}}";
 
@@ -1849,23 +1849,23 @@ namespace System.Text.Json.Serialization.Tests
     public class Parameterized_Class_With_ComplexTuple : ITestClassWithParameterizedCtor
     {
         public Tuple<
-                ClassWithConstructor_SimpleAndComplexParameters,
-                ClassWithConstructor_SimpleAndComplexParameters,
-                ClassWithConstructor_SimpleAndComplexParameters,
-                ClassWithConstructor_SimpleAndComplexParameters,
-                ClassWithConstructor_SimpleAndComplexParameters,
-                ClassWithConstructor_SimpleAndComplexParameters,
-                ClassWithConstructor_SimpleAndComplexParameters> MyTuple { get; }
+                ObjWCtorMixedParams,
+                ObjWCtorMixedParams,
+                ObjWCtorMixedParams,
+                ObjWCtorMixedParams,
+                ObjWCtorMixedParams,
+                ObjWCtorMixedParams,
+                ObjWCtorMixedParams> MyTuple { get; }
 
         public Parameterized_Class_With_ComplexTuple(
             Tuple<
-                ClassWithConstructor_SimpleAndComplexParameters,
-                ClassWithConstructor_SimpleAndComplexParameters,
-                ClassWithConstructor_SimpleAndComplexParameters,
-                ClassWithConstructor_SimpleAndComplexParameters,
-                ClassWithConstructor_SimpleAndComplexParameters,
-                ClassWithConstructor_SimpleAndComplexParameters,
-                ClassWithConstructor_SimpleAndComplexParameters> myTuple) => MyTuple = myTuple;
+                ObjWCtorMixedParams,
+                ObjWCtorMixedParams,
+                ObjWCtorMixedParams,
+                ObjWCtorMixedParams,
+                ObjWCtorMixedParams,
+                ObjWCtorMixedParams,
+                ObjWCtorMixedParams> myTuple) => MyTuple = myTuple;
 
         private const string s_inner_json = @"
             {
